+++ src/components/molecules/GameCard/GameCard.style.ts
